Map update
Here textures and the map were added and tested. The program is now loading a map and selects the appropriate textures to be shown. Only boxes close to the player get updated to increase performance.

In base the imagefilevariables got added.
In Box the whole renderfunctionality got added.
In Map we import a image, we process it into pixels, analyse the pixeldata and create according boxes.
The player got his new collisionsystem to be able to collide with boxes.
Furthermore his Gridposition can now be calculated.
In Utility an importmethod got added for simplicity.
3 texture files got added, those are WIP!